+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ------------------
 - Added customer hasCustomerAccount property
 - Fixed product base price
+- Added new shopware 5.2 attribute management
 
 2.0.4
 -----